# Conflicts:
#	package-lock.json
#	test.js

+++ controllers/users.js
@@ -1,27 +1,27 @@
 require('dotenv').config();
 const { User, Character } = require('../models');
 
-//TODO
-//paranoid세팅으로 임시 삭제이기 때문에 moment(아마) 시간 다루는 api에서 주기적으로 실제 삭제 예정
-//https://runebook.dev/ko/docs/sequelize/manual/paranoid
-const bye = async (req, res) => {
+//paranoid세팅으로 임시 삭제이기 때문에 node-cron에서 주기적으로 실제 삭제
+const bye = async (req, res, next) => {
   try {
+    if (!res.locals.user) return next(myError(401, '로그인되어있지 않습니다'))
     const { id } = res.locals.user;
-    console.log(id)
     User.destroy({ where: { id } })
       .then(() => {
         res.send({ result: true, msg: "회원 탈퇴가 완료되었습니다" })
       })
       .catch((err) => {
-        if (err) throw new Error('db삭제 오류')
+        if (err) return next(new Error('User db삭제 에러'))
       })
   } catch (err) {
-    res.send({ result: false, msg: err.message })
+    console.log(err);
+    return next(err);
   }
 }
 
-const collection = async (req, res) => {
+const collection = async (req, res, next) => {
   try {
+    if (!res.locals.user) return next(myError(401, '로그인되어있지 않습니다'))
     const { id } = res.locals.user;
     const usersCharacter = await Character.findAll({
       where: {
@@ -29,29 +29,33 @@ const collection = async (req, res) => {
         exp: 1000,
       },
     }).catch((err) => {
-      if (err) throw new Error('db검색 오류')
+      console.log(err);
+      if (err) return next(new Error('Character db 검색 에러'));
     })
     return res.send({ result: true, character: usersCharacter, msg: '성공' });
   } catch (err) {
-    return res.send({ result: false, msg: err.message })
+    return next(err);
   }
 }
 
-const updateUser = async (req, res) => {
+const updateUser = async (req, res, next) => {
   try {
     const { id } = res.locals.user;
     if (Object.keys(req.body).length === 0) {
-      throw new Error('수정할 정보가 없습니다.')
+      return next(new Error('수정할 정보가 없습니다.'));
     }
     const { userEmail, nickName, userPw } = req.body;
 
     User.update({ userEmail, nickName, userPw }, { where: { id } })
-      .catch((err) => {
-        if (err) throw new Error('db수정 오류')
+      .then(() => {
+        return res.send({ result: true, msg: '유저 정보가 수정되었습니다' });
       })
-    return res.send({ result: true, msg: '유저 정보가 수정되었습니다' });
+      .catch((err) => {
+        if (err) return next(new Error('유저 정보 수정 db 에러'));
+      })
   } catch (err) {
-    return res.send({ result: false, msg: err.message })
+    console.log(err);
+    return next(err);
   }
 }
